start migration to implicit cleanIntegrationContext

+++ source/server/tests-common.ts
@@ -9,6 +9,7 @@ import chaiAsPromised from "chai-as-promised";
 import sourceMaps from "source-map-support";
 import { AppLocals } from "./utils/locals.js";
 import { StaticConfigShape, parse } from "./utils/config.js";
+import type { Services } from "./create.js";
 import { randomBytes } from "node:crypto";
 import { debuglog } from "node:util";
 sourceMaps.install();
@@ -25,6 +26,7 @@ declare global{
   var resetIntegrationContext :(c:Mocha.Context)=>Promise<void>;
   var getUniqueDb: (name?: string)=>Promise<string>;
   var dropDb: (uri: string)=>Promise<void>;
+  var withIntegrationContext: (config_override?: Record<string, string>) => Services;
 }
 
 global.expect = chai.expect;
@@ -47,8 +49,11 @@ db_uri.password = process.env["PGPASSWORD"]?? "";
 global.getUniqueDb = async function(name?: string){
   console.time("globals getUniqueDb");
   const dbname = name? name.replace(/[^\w]/g, "_").substring(0,58)+"_"+randomBytes(2).toString("hex"): `eCorpus_test_${randomBytes(4).toString("hex")}`;
+  
+  console.time("globals connect");
   const client = new Client({connectionString: new URL(`/postgres`, db_uri).toString()});
   await client.connect();
+  console.timeEnd("globals connect");
   try{
     await client.query(`CREATE DATABASE ${escapeIdentifier(dbname)}`);
   }finally{
@@ -62,8 +67,10 @@ global.getUniqueDb = async function(name?: string){
 
 global.dropDb = async function(uri: string){
   console.time("globals dropDb");
+  console.time("globals connect");
   const client = new Client({connectionString: new URL(`/postgres`, db_uri).toString()});
   await client.connect();
+  console.timeEnd("globals connect");
   try{
     await client.query(`DROP DATABASE ${escapeIdentifier(new URL(uri).pathname.slice(1))} WITH (FORCE)`);
   }finally{
@@ -75,7 +82,7 @@ global.dropDb = async function(uri: string){
 
 global.createIntegrationContext = async function(c :Mocha.Context, config_override :Record<string, string>={}){
   let {default:createService} = await import("./create.js");
-  let titleSlug = "t_"+ (c.currentTest?.title.replace(/[^\w]/g, "_") ?? `eCorpus_integration`)+"_"+randomBytes(4).toString("hex");
+  let titleSlug = "t_"+ (c.currentTest?.title ?? c.test?.parent?.title ?? `eCorpus_integration`).replace(/[^\w]/g, "_").substring(0, 58) +"_"+randomBytes(4).toString("hex");
   c.db_uri = await getUniqueDb(titleSlug);
   c.dir = await fs.mkdtemp(path.join(tmpdir(), titleSlug));
   c.config_env = Object.assign(
@@ -90,13 +97,22 @@ global.createIntegrationContext = async function(c :Mocha.Context, config_overri
   );
   c.services = await createService( c.config_env );
   c.server = c.services.app;
+  const suite = c.test?.parent as any;
+  if(suite?._beforeEach?.includes(c.test)){
+    suite.afterEach(async function(this: Mocha.Context){ await cleanIntegrationContext(this); });
+  } else {
+    suite?.afterAll(async function(this: Mocha.Context){ await cleanIntegrationContext(this); });
+  }
   return c.server.locals;
 }
 
+
 global.cleanIntegrationContext = async function(c :Mocha.Context){
+  if(!c.services) return debuglog("pg:debug")(`No integration context (double close?): ${c.currentTest?.title?? "anonymous"}`);
   await c.services?.close();
   await dropDb(c.db_uri);
   if(c.dir) await fs.rm(c.dir, {recursive: true});
+  delete c.services;
 }
 
 global.resetIntegrationContext = async function(c :Mocha.Context){
